Bundled Arduino Detection and bundled serial communication into arduinoTools.py
