+++ ProjectBuzzer.py
@@ -3,6 +3,7 @@ import sys
 import serial.tools.list_ports
 import time
 import keyboard
+from arduinoTools import findArduino, getButtonStatus
 from playsound import playsound
 from configuration import *
 from tkinter import *
@@ -11,7 +12,7 @@ from tkinter.scrolledtext import ScrolledText
 
 #-------------------------------------------------------------------------
 
-#Global logic variables
+#Private logic variables
 pressedButtonLogger = []
 listOfNum = ["1","2","3", "4", "5", "6", "7", "8"]
 scoreList = [0,0,0,0,0,0,0,0]
@@ -25,22 +26,11 @@ status= ""
 isBorder = True
 #Delay from reset
 resetDelay = 0.1
+winLoseLock = True
+
 #-------------------------------------------------------------------------
 #detect arduino in windows port
-ports = list(serial.tools.list_ports.comports())
-for p in ports:
-    print(p)
-    if "Arduino" in p[1]:
-        arduinoPort = p[0];
-    elif "USB-SERIAL CH340" in p[1]:
-        arduinoPort = p[0];
-#if arduino is found, link arduino serial to python
-if arduinoPort != "":
-    arduino = serial.Serial(arduinoPort, 9600, timeout=0)
-else:
-    print("Error 404, Arduino Not Found!")
-    sys.exit()
-
+arduino = findArduino()
 #-------------------------------------------------------------------------
 
 #Set names on display
@@ -189,39 +179,11 @@ def loseFunc():
     
 #Change win/lose button back to disabled
 def resetButtons():
+    global winLoseLock
     winButton['state'] = 'disabled'
     loseButton['state'] = 'disabled'
+    winLoseLock = True
 
-#-------------------------------------------------------------------------
-#Arduino Communication Functions
-
-#gets latest update from arduino
-def getLatestStatus(): 
-    global status
-    while arduino.inWaiting() > 0:
-        status = arduino.readline();
-    return status;
-
-#strip the raw data of their spaces
-def cleanData(inputData): 
-    if inputData == "": #bug patch between byte and str
-        return "";
-    else:
-        return inputData.strip().decode()
-
-#checks if the data has numbers in it from the input data
-def checkListNum(data):
-    for n in listOfNum:
-        if n in data:
-            return True;
-    return False;
-
-#extract out the numbers from the string
-def extractNum(data):
-    for n in listOfNum:
-        if n in data:
-            return n;
-    return "";
 #-------------------------------------------------------------------------
 #Quit program
 def quit():
@@ -246,11 +208,11 @@ def debugKeyInputs():
     global buttonNumb
     #Keyboard inputs for Reset and Quit
     try:
-        if keyboard.is_pressed('k'): #reset button pressed
+        if keyboard.is_pressed('k') and not winLoseLock: #reset button pressed
             winFunc()
         else:
             pass
-        if keyboard.is_pressed('l'): #reset button pressed
+        if keyboard.is_pressed('l') and not winLoseLock: #reset button pressed
             loseFunc()
         else:
             pass
@@ -330,7 +292,6 @@ def updateScoreBoard(points):
     global currentPlayer
     global scoreList
     global scoreDisplay
-    global listOfNum
     scoreList[currentPlayer] += points
     scoreDisplay[currentPlayer].set(
         str(listOfNum[currentPlayer]) + "\n" + str(scoreList[currentPlayer]))
@@ -338,19 +299,15 @@ def updateScoreBoard(points):
 #-------------------------------------------------------------------------
 def main():
     global unlocked
-    global listOfNum
     global currentName
     global winButton
     global loseButton
-    global root
     global buttonNumb
     global pressedButtonLogger
-    global debugging
     global currentPlayer
-    global selectedColor
-    global scoreConfig
+    global winLoseLock
     #Assign values from arduino serial output
-    buttonNumb = extractNum(cleanData(getLatestStatus())) #Decode arduino output from \r\n'1'
+    buttonNumb = getButtonStatus(arduino) #Decode arduino output from \r\n'1'
     if debugging:
         debugKeyInputs()
     if buttonNumb != "" and unlocked and checkButtonNotPressedBefore(buttonNumb):
@@ -360,6 +317,7 @@ def main():
         playsound('sound/ButtonPressed.wav', False)
         #Lock button
         unlocked = False
+        winLoseLock = False
         #Set name selected from nameList, from arduino output
         currentName = nameList[int(buttonNumb)-1]
         #Set current player to press button
